+++ test/BaseListBoxTest.ts
@@ -3,6 +3,9 @@
 
 import TestHelper = require("./infrastructure/TestHelper");
 import ListboxSettings = require("../src/ts/contract/ListboxSettings");
+import ListboxEvent = require("../src/ts/event/ListboxEvent");
+
+/* tslint:disable:no-string-literal */
 
 QUnit.module( "BaseListBoxTest", {
     beforeEach: function(): void {
@@ -549,3 +552,80 @@ QUnit.test("check getSelection", function (): void {
     QUnit.assert.equal(selection[0].id, "id02");
     QUnit.assert.equal(selection[1].id, "subid01");
 });
+
+QUnit.test("check valueChanged event", function (): void {
+    var target: JQuery = null;
+
+    var delegate: any = (event: ListboxEvent) => {
+        QUnit.assert.equal(event.eventName, ListboxEvent.VALUE_CHANGED);
+        QUnit.assert.equal(event.target, target);
+        QUnit.assert.equal(event.args, "mySpecialValue");
+    };
+
+    var listbox: ExtendedListboxInstance = TestHelper.generateSingleList({ onValueChanged: delegate });
+    target = listbox.target;
+
+    listbox["listbox"].baseListBox.fireEvent(ListboxEvent.VALUE_CHANGED, "mySpecialValue");
+});
+
+QUnit.test("check itemsChanged event", function (): void {
+    var target: JQuery = null;
+
+    var delegate: any = (event: ListboxEvent) => {
+        QUnit.assert.equal(event.eventName, ListboxEvent.ITEMS_CHANGED);
+        QUnit.assert.equal(event.target, target);
+        QUnit.assert.equal(event.args, "mySpecialValue");
+    };
+
+    var listbox: ExtendedListboxInstance = TestHelper.generateSingleList({ onItemsChanged: delegate });
+    target = listbox.target;
+
+    listbox["listbox"].baseListBox.fireEvent(ListboxEvent.ITEMS_CHANGED, "mySpecialValue");
+});
+
+QUnit.test("check filterChanged event", function (): void {
+    var target: JQuery = null;
+
+    var delegate: any = (event: ListboxEvent) => {
+        QUnit.assert.equal(event.eventName, ListboxEvent.FILTER_CHANGED);
+        QUnit.assert.equal(event.target, target);
+        QUnit.assert.equal(event.args, "mySpecialValue");
+    };
+
+    var listbox: ExtendedListboxInstance = TestHelper.generateSingleList({ onFilterChanged: delegate });
+    target = listbox.target;
+
+    listbox["listbox"].baseListBox.fireEvent(ListboxEvent.FILTER_CHANGED, "mySpecialValue");
+});
+
+QUnit.test("check itemEnterPressed event", function (): void {
+    var target: JQuery = null;
+
+    var delegate: any = (event: ListboxEvent) => {
+        QUnit.assert.equal(event.eventName, ListboxEvent.ITEM_ENTER_PRESSED);
+        QUnit.assert.equal(event.target, target);
+        QUnit.assert.equal(event.args, "mySpecialValue");
+    };
+
+    var listbox: ExtendedListboxInstance = TestHelper.generateSingleList({ onItemEnterPressed: delegate });
+    target = listbox.target;
+
+    listbox["listbox"].baseListBox.fireEvent(ListboxEvent.ITEM_ENTER_PRESSED, "mySpecialValue");
+});
+
+QUnit.test("check itemDoubleClicked event", function (): void {
+    var target: JQuery = null;
+
+    var delegate: any = (event: ListboxEvent) => {
+        QUnit.assert.equal(event.eventName, ListboxEvent.ITEM_DOUBLE_CLICKED);
+        QUnit.assert.equal(event.target, target);
+        QUnit.assert.equal(event.args, "mySpecialValue");
+    };
+
+    var listbox: ExtendedListboxInstance = TestHelper.generateSingleList({ onItemDoubleClicked: delegate });
+    target = listbox.target;
+
+    listbox["listbox"].baseListBox.fireEvent(ListboxEvent.ITEM_DOUBLE_CLICKED, "mySpecialValue");
+});
+
+/* tslint:enable:no-string-literal */
